Minor: Extract parquet row group pruning code into its own module
